// src/app/models/model/product/barcodeModel.ts
export class BarcodeModel {
  barcode: string;
  itemCode: string;
  colorCode: string;
  itemDim1Code: string;

  itemDescription: string;
  warehouse: string;
  shelfNo: string;
  party: string;
  attributeCode: string;
  priceCurrency: string;
  price: number;
  quantity: number;
  picture: string;
  picture2: string;
}